Renamed BusTest to ElectriCityBusTest and removed old unused tests
